Add essential folder to keep folder structure
When including only the icons and the logos, we need to reference a
folder (otherwise we would have to change it everywhere). This way we
can keep the `{{inline-svg "icons/alert"}}` if we include
`node_modules/ember-cli-uniq/assets/images/essential/` in the
`inline-svg` paths setting.

This is a breaking change (even though it is easy upgradable) we must
consider this